CEMS-2065 added usecase test which was previously only covered by an api test, moved some methods so they are callable by other classes who need them
